+++ aws-opentelemetry-distro/tests/amazon/opentelemetry/distro/patches/test_instrumentation_patch.py
@@ -34,7 +34,7 @@ from amazon.opentelemetry.distro.semconv._incubating.attributes.gen_ai_attribute
     GEN_AI_RUNTIME_ID,
 )
 from opentelemetry.instrumentation.botocore import BotocoreInstrumentor
-from opentelemetry.instrumentation.botocore.extensions import _KNOWN_EXTENSIONS, bedrock_utils
+from opentelemetry.instrumentation.botocore.extensions import _BOTOCORE_EXTENSIONS, bedrock_utils
 from opentelemetry.propagate import get_global_textmap
 from opentelemetry.semconv._incubating.attributes.aws_attributes import (
     AWS_BEDROCK_GUARDRAIL_ID,
@@ -178,55 +178,58 @@ class TestInstrumentationPatch(TestCase):
 
     def _test_unpatched_botocore_instrumentation(self):
         # Kinesis
-        self.assertFalse("kinesis" in _KNOWN_EXTENSIONS, "Upstream has added a Kinesis extension")
+        self.assertFalse("kinesis" in _BOTOCORE_EXTENSIONS, "Upstream has added a Kinesis extension")
 
         # S3
-        self.assertFalse("s3" in _KNOWN_EXTENSIONS, "Upstream has added a S3 extension")
+        self.assertFalse("s3" in _BOTOCORE_EXTENSIONS, "Upstream has added a S3 extension")
 
         # SQS
-        self.assertTrue("sqs" in _KNOWN_EXTENSIONS, "Upstream has removed the SQS extension")
+        self.assertTrue("sqs" in _BOTOCORE_EXTENSIONS, "Upstream has removed the SQS extension")
 
         # Bedrock
-        self.assertFalse("bedrock" in _KNOWN_EXTENSIONS, "Upstream has added a Bedrock extension")
+        self.assertFalse("bedrock" in _BOTOCORE_EXTENSIONS, "Upstream has added a Bedrock extension")
 
         # Bedrock Agent
-        self.assertFalse("bedrock-agent" in _KNOWN_EXTENSIONS, "Upstream has added a Bedrock Agent extension")
+        self.assertFalse("bedrock-agent" in _BOTOCORE_EXTENSIONS, "Upstream has added a Bedrock Agent extension")
 
         # Bedrock Agent Runtime
         self.assertFalse(
-            "bedrock-agent-runtime" in _KNOWN_EXTENSIONS, "Upstream has added a Bedrock Agent Runtime extension"
+            "bedrock-agent-runtime" in _BOTOCORE_EXTENSIONS, "Upstream has added a Bedrock Agent Runtime extension"
         )
 
         # Bedrock AgentCore
-        self.assertFalse("bedrock-agentcore" in _KNOWN_EXTENSIONS, "Upstream has added a Bedrock AgentCore extension")
+        self.assertFalse(
+            "bedrock-agentcore" in _BOTOCORE_EXTENSIONS, "Upstream has added a Bedrock AgentCore extension"
+        )
 
         # Bedrock AgentCore Control
         self.assertFalse(
-            "bedrock-agentcore-control" in _KNOWN_EXTENSIONS, "Upstream has added a Bedrock AgentCore Control extension"
+            "bedrock-agentcore-control" in _BOTOCORE_EXTENSIONS,
+            "Upstream has added a Bedrock AgentCore Control extension",
         )
 
         # BedrockRuntime
-        self.assertTrue("bedrock-runtime" in _KNOWN_EXTENSIONS, "Upstream has added a bedrock-runtime extension")
+        self.assertTrue("bedrock-runtime" in _BOTOCORE_EXTENSIONS, "Upstream has added a bedrock-runtime extension")
 
         # SecretsManager
-        self.assertTrue("secretsmanager" in _KNOWN_EXTENSIONS, "Upstream has removed the SecretsManager extension")
+        self.assertTrue("secretsmanager" in _BOTOCORE_EXTENSIONS, "Upstream has removed the SecretsManager extension")
 
         # SNS
-        self.assertTrue("sns" in _KNOWN_EXTENSIONS, "Upstream has removed the SNS extension")
+        self.assertTrue("sns" in _BOTOCORE_EXTENSIONS, "Upstream has removed the SNS extension")
 
         # StepFunctions
-        self.assertTrue("stepfunctions" in _KNOWN_EXTENSIONS, "Upstream has removed the StepFunctions extension")
+        self.assertTrue("stepfunctions" in _BOTOCORE_EXTENSIONS, "Upstream has removed the StepFunctions extension")
 
         # Lambda
-        self.assertTrue("lambda" in _KNOWN_EXTENSIONS, "Upstream has removed the Lambda extension")
+        self.assertTrue("lambda" in _BOTOCORE_EXTENSIONS, "Upstream has removed the Lambda extension")
 
         # DynamoDB
-        self.assertTrue("dynamodb" in _KNOWN_EXTENSIONS, "Upstream has removed a DynamoDB extension")
+        self.assertTrue("dynamodb" in _BOTOCORE_EXTENSIONS, "Upstream has removed a DynamoDB extension")
 
     # pylint: disable=too-many-statements, too-many-locals
     def _test_patched_botocore_instrumentation(self):
         # Kinesis
-        self.assertTrue("kinesis" in _KNOWN_EXTENSIONS)
+        self.assertTrue("kinesis" in _BOTOCORE_EXTENSIONS)
         kinesis_attributes: Dict[str, str] = _do_extract_kinesis_attributes()
         self.assertTrue(AWS_KINESIS_STREAM_NAME in kinesis_attributes)
         self.assertEqual(kinesis_attributes[AWS_KINESIS_STREAM_NAME], _STREAM_NAME)
@@ -234,13 +237,13 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(kinesis_attributes[AWS_KINESIS_STREAM_ARN], _STREAM_ARN)
 
         # S3
-        self.assertTrue("s3" in _KNOWN_EXTENSIONS)
+        self.assertTrue("s3" in _BOTOCORE_EXTENSIONS)
         s3_attributes: Dict[str, str] = _do_extract_s3_attributes()
         self.assertTrue(AWS_S3_BUCKET in s3_attributes)
         self.assertEqual(s3_attributes[AWS_S3_BUCKET], _BUCKET_NAME)
 
         # SQS
-        self.assertTrue("sqs" in _KNOWN_EXTENSIONS)
+        self.assertTrue("sqs" in _BOTOCORE_EXTENSIONS)
         sqs_attributes: Dict[str, str] = _do_extract_sqs_attributes()
         self.assertTrue(AWS_SQS_QUEUE_URL in sqs_attributes)
         self.assertEqual(sqs_attributes[AWS_SQS_QUEUE_URL], _QUEUE_URL)
@@ -264,7 +267,7 @@ class TestInstrumentationPatch(TestCase):
         self._test_patched_from_converse_with_malformed_response()
 
         # Bedrock Agent Runtime
-        self.assertTrue("bedrock-agent-runtime" in _KNOWN_EXTENSIONS)
+        self.assertTrue("bedrock-agent-runtime" in _BOTOCORE_EXTENSIONS)
         bedrock_agent_runtime_attributes: Dict[str, str] = _do_extract_attributes_bedrock("bedrock-agent-runtime")
         self.assertEqual(len(bedrock_agent_runtime_attributes), 2)
         self.assertEqual(bedrock_agent_runtime_attributes[AWS_BEDROCK_AGENT_ID], _BEDROCK_AGENT_ID)
@@ -273,8 +276,8 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(len(bedrock_agent_runtime_sucess_attributes), 0)
 
         # Bedrock AgentCore
-        self.assertTrue("bedrock-agentcore" in _KNOWN_EXTENSIONS)
-        self.assertTrue("bedrock-agentcore-control" in _KNOWN_EXTENSIONS)
+        self.assertTrue("bedrock-agentcore" in _BOTOCORE_EXTENSIONS)
+        self.assertTrue("bedrock-agentcore-control" in _BOTOCORE_EXTENSIONS)
 
         _do_extract_bedrock_agentcore_attributes, _do_on_success_bedrock_agentcore = _do_bedrock_agentcore_tests()
         bedrock_agentcore_attributes: Dict[str, str] = _do_extract_bedrock_agentcore_attributes()
@@ -314,10 +317,10 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(name_success_attrs[AWS_AUTH_CREDENTIAL_PROVIDER], _AGENTCORE_CREDENTIAL_PROVIDER_NAME)
 
         # BedrockRuntime
-        self.assertTrue("bedrock-runtime" in _KNOWN_EXTENSIONS)
+        self.assertTrue("bedrock-runtime" in _BOTOCORE_EXTENSIONS)
 
         # SecretsManager
-        self.assertTrue("secretsmanager" in _KNOWN_EXTENSIONS)
+        self.assertTrue("secretsmanager" in _BOTOCORE_EXTENSIONS)
         secretsmanager_attributes: Dict[str, str] = _do_extract_secretsmanager_attributes()
         self.assertTrue(AWS_SECRETSMANAGER_SECRET_ARN in secretsmanager_attributes)
         self.assertEqual(secretsmanager_attributes[AWS_SECRETSMANAGER_SECRET_ARN], _SECRET_ARN)
@@ -326,13 +329,13 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(secretsmanager_success_attributes[AWS_SECRETSMANAGER_SECRET_ARN], _SECRET_ARN)
 
         # SNS
-        self.assertTrue("sns" in _KNOWN_EXTENSIONS)
+        self.assertTrue("sns" in _BOTOCORE_EXTENSIONS)
         sns_attributes: Dict[str, str] = _do_extract_sns_attributes()
         self.assertTrue(AWS_SNS_TOPIC_ARN in sns_attributes)
         self.assertEqual(sns_attributes[AWS_SNS_TOPIC_ARN], _TOPIC_ARN)
 
         # StepFunctions
-        self.assertTrue("stepfunctions" in _KNOWN_EXTENSIONS)
+        self.assertTrue("stepfunctions" in _BOTOCORE_EXTENSIONS)
         stepfunctions_attributes: Dict[str, str] = _do_extract_stepfunctions_attributes()
         self.assertTrue(AWS_STEP_FUNCTIONS_STATE_MACHINE_ARN in stepfunctions_attributes)
         self.assertEqual(stepfunctions_attributes[AWS_STEP_FUNCTIONS_STATE_MACHINE_ARN], _STATE_MACHINE_ARN)
@@ -340,7 +343,7 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(stepfunctions_attributes[AWS_STEP_FUNCTIONS_ACTIVITY_ARN], _ACTIVITY_ARN)
 
         # Lambda
-        self.assertTrue("lambda" in _KNOWN_EXTENSIONS)
+        self.assertTrue("lambda" in _BOTOCORE_EXTENSIONS)
         lambda_attributes: Dict[str, str] = _do_extract_lambda_attributes()
         self.assertTrue(AWS_LAMBDA_FUNCTION_NAME in lambda_attributes)
         self.assertEqual(lambda_attributes[AWS_LAMBDA_FUNCTION_NAME], _LAMBDA_FUNCTION_NAME)
@@ -348,7 +351,7 @@ class TestInstrumentationPatch(TestCase):
         self.assertEqual(lambda_attributes[AWS_LAMBDA_RESOURCE_MAPPING_ID], _LAMBDA_SOURCE_MAPPING_ID)
 
         # DynamoDB
-        self.assertTrue("dynamodb" in _KNOWN_EXTENSIONS)
+        self.assertTrue("dynamodb" in _BOTOCORE_EXTENSIONS)
         dynamodb_success_attributes: Dict[str, str] = _do_on_success_dynamodb()
         self.assertTrue(AWS_DYNAMODB_TABLE_ARN in dynamodb_success_attributes)
         self.assertEqual(dynamodb_success_attributes[AWS_DYNAMODB_TABLE_ARN], _TABLE_ARN)
@@ -388,16 +391,22 @@ class TestInstrumentationPatch(TestCase):
 
         with patch(
             "opentelemetry.instrumentation.botocore._determine_call_context", return_value=mock_call_context
-        ), patch("opentelemetry.instrumentation.botocore._find_extension", return_value=mock_extension), patch(
+        ), patch(
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_extension",
+            return_value=mock_extension,
+        ), patch(
             "opentelemetry.instrumentation.botocore.is_instrumentation_enabled", return_value=True
         ), patch(
             "amazon.opentelemetry.distro.patches._botocore_patches.get_server_attributes", return_value={}
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_tracer", return_value=mock_tracer
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_tracer",
+            return_value=mock_tracer,
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_logger", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_logger",
+            return_value=MagicMock(),
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_meter", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_metrics",
+            return_value={},
         ):
             instrumentor = BotocoreInstrumentor()
             instrumentor.instrument()
@@ -436,16 +445,22 @@ class TestInstrumentationPatch(TestCase):
 
         with patch(
             "opentelemetry.instrumentation.botocore._determine_call_context", return_value=mock_call_context
-        ), patch("opentelemetry.instrumentation.botocore._find_extension", return_value=mock_extension), patch(
+        ), patch(
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_extension",
+            return_value=mock_extension,
+        ), patch(
             "opentelemetry.instrumentation.botocore.is_instrumentation_enabled", return_value=True
         ), patch(
             "amazon.opentelemetry.distro.patches._botocore_patches.get_server_attributes", return_value={}
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_tracer", return_value=mock_tracer
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_tracer",
+            return_value=mock_tracer,
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_logger", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_logger",
+            return_value=MagicMock(),
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_meter", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_metrics",
+            return_value={},
         ):
             instrumentor = BotocoreInstrumentor()
             instrumentor.instrument()
@@ -484,16 +499,22 @@ class TestInstrumentationPatch(TestCase):
 
         with patch(
             "opentelemetry.instrumentation.botocore._determine_call_context", return_value=mock_call_context
-        ), patch("opentelemetry.instrumentation.botocore._find_extension", return_value=mock_extension), patch(
+        ), patch(
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_extension",
+            return_value=mock_extension,
+        ), patch(
             "opentelemetry.instrumentation.botocore.is_instrumentation_enabled", return_value=True
         ), patch(
             "amazon.opentelemetry.distro.patches._botocore_patches.get_server_attributes", return_value={}
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_tracer", return_value=mock_tracer
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_tracer",
+            return_value=mock_tracer,
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_logger", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_logger",
+            return_value=MagicMock(),
         ), patch(
-            "opentelemetry.instrumentation.botocore.get_meter", return_value=MagicMock()
+            "opentelemetry.instrumentation.botocore.extensions.registry.ExtensionRegistry.get_metrics",
+            return_value={},
         ):
             instrumentor = BotocoreInstrumentor()
             instrumentor.instrument()
@@ -745,7 +766,7 @@ class TestInstrumentationPatch(TestCase):
     def _test_patched_bedrock_agent_instrumentation(self):
         """For bedrock-agent service, both extract_attributes and on_success provides attributes,
         the attributes depend on the API being invoked."""
-        self.assertTrue("bedrock-agent" in _KNOWN_EXTENSIONS)
+        self.assertTrue("bedrock-agent" in _BOTOCORE_EXTENSIONS)
         operation_to_expected_attribute = {
             "CreateAgentActionGroup": (AWS_BEDROCK_AGENT_ID, _BEDROCK_AGENT_ID),
             "CreateAgentAlias": (AWS_BEDROCK_AGENT_ID, _BEDROCK_AGENT_ID),
@@ -1286,7 +1307,7 @@ def _do_extract_attributes(service_name: str, params: Dict[str, Any], operation:
     if operation:
         mock_call_context.operation = operation
     attributes: Dict[str, str] = {}
-    sqs_extension = _KNOWN_EXTENSIONS[service_name]()(mock_call_context)
+    sqs_extension = _BOTOCORE_EXTENSIONS[service_name]()(mock_call_context)
     sqs_extension.extract_attributes(attributes)
     return attributes
 
@@ -1316,7 +1337,7 @@ def _do_on_success(
     if params:
         mock_call_context.params = params
 
-    extension = _KNOWN_EXTENSIONS[service_name]()(mock_call_context)
+    extension = _BOTOCORE_EXTENSIONS[service_name]()(mock_call_context)
     extension.on_success(span_mock, result, mock_instrumentor_context)
 
     return span_attributes